backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
google-cloud-firestore==2.13.1
python-dotenv==1.0.0
numpy==1.24.3
torch==2.1.0
transformers==4.35.0
